remainder of @cpgr suggestions implemented
